// src/models/f1/car-frame.ts
export const CAR_FRAME = {
// ...
  parts: {
    "front-wing": {
      FrontWingLivery: {
        min: [-3.012, 0.129, -0.595],
        max: [-2.322, 0.309, 0.595]
      },
      material_0: {
        min: [-2.578, 0.283, -0.878],
        max: [-2.339, 0.33, 0.878]
      },
      FrontEndplateLivery: {
        min: [-2.759, 0.104, -0.882],
        max: [-2.094, 0.394, 0.882]
      },
      Carbon: {
        min: [-2.88, 0.142, -0.004],
        max: [-2.696, 0.183, 0.004]
      }
    },
    "rear-wing": {
      RearWingLivery: {
        min: [1.943, 0.815, -0.566],
        max: [2.327, 0.935, 0.566]
      },
      RearEndplateLivery: {
        min: [1.932, 0.369, -0.577],
        max: [2.456, 0.936, 0.577]
      },
      Carbon: {
        min: [1.793, 0.434, -0.07],
        max: [2.249, 0.964, 0.07]
      }
    },
// ...
  }
} as const;